Refactor code to reuse existing classes
