+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/LockFilePackage.java
@@ -1,15 +1,19 @@
 package org.wso2.ballerinalang.compiler;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Class to handle packages in Ballerina.lock.
+ *
+ * @since 0.973.1
  */
 public class LockFilePackage {
-    private String org;
-    private String name;
-    private String version;
-    private List<LockFilePackage> dependencyPackages;
+    private String org = "";
+    private String name = "";
+    private String version = "";
+    private List<LockFilePackage> dependencyPackages = new ArrayList<>();
 
     /**
      * Constructor.
@@ -25,6 +29,12 @@ public class LockFilePackage {
     }
 
     /**
+     * Default constructor.
+     */
+    public LockFilePackage() {
+    }
+
+    /**
      * Get package name.
      *
      * @return package name
@@ -34,12 +44,30 @@ public class LockFilePackage {
     }
 
     /**
+     * Set package name.
+     *
+     * @param name package name
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    /**
      * Get package version.
      *
      * @return version
      */
     public String getVersion() {
         return version;
+    }
+
+    /**
+     * Set package version.
+     *
+     * @param version version
+     */
+    public void setVersion(String version) {
+        this.version = version;
     }
 
     /**
@@ -67,5 +95,34 @@ public class LockFilePackage {
      */
     public String getOrg() {
         return org;
+    }
+
+    /**
+     * Set the org-name of the package.
+     *
+     * @param org org-name of the package
+     */
+    public void setOrg(String org) {
+        this.org = org;
+    }
+
+    /**
+     * Remove duplicates from import packages list.
+     *
+     * @param list list of elements
+     * @return import packages list without duplicates
+     */
+    private List<LockFilePackage> removeDuplicates(List<LockFilePackage> list) {
+        return list.stream().distinct().collect(Collectors.toList());
+    }
+
+    /**
+     * Add a package to the imports list.
+     *
+     * @param lockFilePackage imported package
+     */
+    public void addImport(LockFilePackage lockFilePackage) {
+        this.dependencyPackages.add(lockFilePackage);
+        dependencyPackages = removeDuplicates(dependencyPackages);
     }
 }
